Implemented a "WebDAV" CloudProvider

+++ src/main/java/org/cryptomator/cloudaccess/api/ProgressListener.java
@@ -4,6 +4,6 @@ public interface ProgressListener {
 
 	ProgressListener NO_PROGRESS_AWARE = value -> {};
 
-	void onProgress(int value);
+	void onProgress(long value);
 
 }
